feat(themes/publish): Use the Mojang session server to validate requests
The client sends us a username and a shared secret in the body of the
request, we then combine this hash with a shared constant to get the
actual server ID hash that was sent to the login endpoint.

We can then use this hash to ensure that the request has at least came
from an actual Minecraft account, and not someone trying to spam our
API with bogus requests.